+++ packages/rx_bloc_generator/test/src/rx_bloc_test_generic_bloc.dart
@@ -1,4 +1,5 @@
 @Tags(['not-tests'])
+library;
 
 import 'package:rx_bloc/rx_bloc.dart';
 import 'package:source_gen_test/source_gen_test.dart';