+++ packages/rx_bloc_generator/test/src/rx_bloc_test_seed_for_publish_subject_error.dart
@@ -1,4 +1,5 @@
 @Tags(['not-tests'])
+library;
 
 import 'package:rx_bloc/rx_bloc.dart';
 // ignore: import_of_legacy_library_into_null_safe